cleanup and fix tests
