+++ src/ForwardDiffChainRules.jl
@@ -33,23 +33,7 @@ module ForwardDiffChainRules
 using ChainRulesCore
 import ForwardDiff
 import MacroTools
-using NamedTupleTools
-
-using Requires
-
-include("flatten.jl")
-
-function __init__()
-    @require SparseArrays="2f01184e-e22b-5df5-ae63-d93ebab69eaf" begin
-        using .SparseArrays: SparseMatrixCSCUnion, SparseVectorUnion, getnzval, getrowval, getcolptr
-        include("flatten_SparseArrays.jl")
-    end
-
-    @require JuMP="4076af6c-e467-56ae-b986-b466b2749572" begin
-        import .JuMP
-        include("flatten_JuMP.jl")
-    end
-end
+using DifferentiableFlatten
 
 macro ForwardDiff_frule(sig)
     _fd_frule(sig)
@@ -62,23 +46,22 @@ function _fd_frule(sig)
         function $(esc(f))($(esc.(x)...))
             f = $(esc(f))
             x = ($(esc.(x)...),)
-            flatx, unflattenx = ForwardDiffChainRules.flatten(x)
+            flatx, unflattenx = ForwardDiffChainRules.DifferentiableFlatten.flatten(x)
             CS = length(ForwardDiff.partials(first(flatx)))
             flat_xprimals = ForwardDiff.value.(flatx)
             flat_xpartials = reduce(vcat, transpose.(ForwardDiff.partials.(flatx)))
 
             xprimals = unflattenx(flat_xprimals)
             xpartials1 = unflattenx(flat_xpartials[:,1])
-
             yprimals, ypartials1 = ChainRulesCore.frule(
                 (NoTangent(), xpartials1...), f, xprimals...,
             )
-            flat_yprimals, unflatteny = ForwardDiffChainRules.flatten(yprimals)
-            flat_ypartials1, _ = ForwardDiffChainRules.flatten(ypartials1)
+            flat_yprimals, unflatteny = ForwardDiffChainRules.DifferentiableFlatten.flatten(yprimals)
+            flat_ypartials1, _ = ForwardDiffChainRules.DifferentiableFlatten.flatten(ypartials1)
             flat_ypartials = hcat(reshape(flat_ypartials1, :, 1), ntuple(Val(CS - 1)) do i
                 xpartialsi = unflattenx(flat_xpartials[:, i+1])
                 _, ypartialsi = ChainRulesCore.frule((NoTangent(), xpartialsi...), f, xprimals...)
-                return ForwardDiffChainRules.flatten(ypartialsi)[1]
+                return ForwardDiffChainRules.DifferentiableFlatten.flatten(ypartialsi)[1]
             end...)
 
             T = ForwardDiff.tagtype(eltype(flatx))
@@ -88,19 +71,6 @@ function _fd_frule(sig)
             return unflatteny(flaty)
         end
     end
-end
-
-function flatten(::Nothing)
-    return Float64[], _ -> nothing
-end
-function flatten(::NoTangent)
-    return Float64[], _ -> NoTangent()
-end
-function flatten(::ZeroTangent)
-    return Float64[], _ -> ZeroTangent()
-end
-function flatten(::Tuple{})
-    return Float64[], _ -> ()
 end
 
 end # module ForwardDiffChainRules
